myapp/pyproject.toml
[tool.poetry]
name = "myapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
PyJWT = "2.9.0"
setuptools = "74.0.0"
sqlparse = "0.5.1"
wheel = "0.44.0"
psycopg2-binary = "2.9.10"
django = "^5.1.3"

[tool.poetry.dev-dependencies]
pytest = "8.3.3"
coverage = "7.6.4"
faker = "^30.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
